Restructure AMQP protocol definitions
